requirements.txt
# Add required packages here!
requests~=2.31.0
python-dotenv~=1.0.1
urllib3~=2.2.1